requirements.txt
typer
numpy
tqdm
wasabi
setuptools
wheel
spacy-experimental
sentence_transformers